1.3 release

+++ CHGKBet/Views/Preparation/PreparationPresenter.swift
@@ -3,7 +3,8 @@ import RxSwift
 
 class PreparationPresenter {
     
-    private let service = UsersService()
+    private let usersService = UsersService()
+    private let appInfoService = AppInfoService()
     private let disposeBag = DisposeBag()
 
     private weak var preparationViewController: PreparationViewController?
@@ -11,26 +12,71 @@ class PreparationPresenter {
     init(preparationViewController: PreparationViewController) {
         self.preparationViewController = preparationViewController
     }
-    
-    func checkUser() {
-        if let credentials = User.getCredentials() {
-            service.requestUserInfo(credentials: credentials)
-                .subsribe { result in
-                    switch result {
-                    case .failure(_):
-                        KeychainService.removePassword(service: KeychainService.authService)
+
+    func checkConditions() {
+        let allChecks = Completable.merge([checkVersion(), checkUser()])
+        allChecks.subscribe(
+            onCompleted: {
+                LinePresenter.present()
+            },
+            onError: { error in
+                if let appError = error as? AppError {
+                    switch appError {
+                    case .invalidCredentials:
                         AuthorizationPresenter.present()
-                    case .success(let userInfo):
-                        User.currentUser = User(credentials: credentials, info: userInfo)
-                        LinePresenter.present()
-                    case .none:
-                        return
+                    case .versionLessThanMinimum:
+                        InsufficientVersionPresenter.present()
+                    default:
+                        break
                     }
                 }
-                .disposed(by: disposeBag)
-        } else {
-            KeychainService.removePassword(service: KeychainService.authService)
-            AuthorizationPresenter.present()
+            }
+        )
+        .disposed(by: disposeBag)
+    }
+
+    private func checkVersion() -> Completable {
+        Completable.create { [weak self] observer -> Disposable in
+            guard let self = self else { return Disposables.create {} }
+            let userVersion = AppVersion(Bundle.main.version)!
+            return self.appInfoService.requestVersions().subscribe(
+                onSuccess: { appVersions in
+                    guard let minimumVersion = appVersions.appVersionMinimum else {
+                        observer(.completed)
+                        return
+                    }
+                    if userVersion >= minimumVersion {
+                        observer(.completed)
+                    } else {
+                        observer(.error(AppError.versionLessThanMinimum))
+                    }
+                }, onError: { error in
+                    observer(.error(error))
+                }
+            )
+        }
+    }
+    
+    private func checkUser() -> Completable {
+        return Completable.create { [weak self] observer -> Disposable in
+            if let credentials = User.getCredentials() {
+                guard let self = self else { return Disposables.create {} }
+                return self.usersService.requestUserInfo(credentials: credentials)
+                    .subscribe(onSuccess: { userInfo in
+                        User.currentUser = User(credentials: credentials, info: userInfo)
+                        observer(.completed)
+                    },
+                    onError: { error in
+                        KeychainService.removePassword(service: KeychainService.authService)
+                        observer(.error(AppError.invalidCredentials))
+                    }
+                )
+            } else {
+                KeychainService.removePassword(service: KeychainService.authService)
+                AuthorizationPresenter.present()
+                observer(.error(AppError.invalidCredentials))
+                return Disposables.create { }
+            }
         }
     }
